+++ bcbio/bam/__init__.py
@@ -254,6 +254,28 @@ def check_header(in_bam, rgnames, ref_file, config):
     _check_bam_contigs(in_bam, ref_file, config)
     _check_sample(in_bam, rgnames)
 
+def has_tags(in_bam, tags):
+    """
+    Checks the first 10,000 records in a BAM file for a set of tags.
+    Returns True if all the tags are seen at least once and False otherwise.
+    """
+    assert isinstance(tags, set), f"{tags} is not a set."
+    assert is_bam(in_bam), f"{in_bam} is not a BAM file."
+    MAX_ALIGNMENTS = 10000
+    with pysam.Samfile(in_bam, "rb") as bamfile:
+        alignments = 0
+        for read in bamfile.fetch(until_eof=True):
+            alignments += 1
+            if alignments > MAX_ALIGNMENTS:
+                return False
+            if not tags:
+                return True
+            tags_left = tags.copy()
+            for tag in tags_left:
+                if read.has_tag(tag):
+                    tags.remove(tag)
+    return False
+
 def _check_sample(in_bam, rgnames):
     """Ensure input sample name matches expected run group names.
     """